control_p2.rs hands its LowerCtx match impl to control_p2_e.rs, a pure text move under the 800-line ceiling
